layout_pipeline/requirements.txt
torch>=2.0.0
transformers>=4.35.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
layoutlmv3>=0.0.1
detectron2>=0.6; platform_system != "Darwin"
opencv-python>=4.8.0
pytesseract>=0.3.10


